Bugfixes & minor improvements

+++ compiler/rust/ppc/src/compiler.rs
@@ -4,6 +4,7 @@ use std::{
 	io,
 	io::prelude::*,
 	io::Error,
+	io::ErrorKind,
 	process::Command,
 	str,
 	usize,
@@ -614,24 +615,24 @@ pub fn parse_statement(tokens: &Vec<Token>, functions: &Vec<Function>, i: &mut u
 					Kind::Var(ref name, _) if depth2 == 0 => if let Some(def) = is_defined(functions, name) {
 						match highest {
 							Some(func) => match func.1 {
-								Some(def2) => if (def.precedence > def2.precedence && depth + depth2 == func.2) || depth + depth2 > func.2 {
-									highest = Some((*i, Some(def), depth + depth2));
+								Some(def2) => if (def.precedence > def2.precedence && depth == func.2) || depth > func.2 {
+									highest = Some((*i, Some(def), depth));
 								},
-								None => if depth + depth2 > func.2 {
-									highest = Some((*i, Some(def), depth + depth2));
+								None => if depth >= func.2 {
+									highest = Some((*i, Some(def), depth));
 								}
 							},
-							None => highest = Some((*i, Some(def), depth + depth2))
+							None => highest = Some((*i, Some(def), depth))
 						}
 					},
 					
 					Kind::GroupOp(ref op) if op == "{" => {
 						if depth2 == 0 {
 							match highest {
-								Some(func) => if depth + depth2 >= func.2 {
-									highest = Some((*i, None, depth + depth2));
+								Some(func) => if depth >= func.2 {
+									highest = Some((*i, None, depth));
 								},
-								None => highest = Some((*i, None, depth + depth2))
+								None => highest = Some((*i, None, depth))
 							}
 						}
 						
@@ -665,14 +666,14 @@ pub fn parse_statement(tokens: &Vec<Token>, functions: &Vec<Function>, i: &mut u
 							if let Some(def) = is_defined(functions, &name) {
 								match highest {
 									Some(func) => match func.1 {
-										Some(def2) => if (def.precedence > def2.precedence && depth + depth2 == func.2) || depth + depth2 > func.2 {
-											highest = Some((start, Some(def), depth + depth2));
+										Some(def2) => if (def.precedence > def2.precedence && depth == func.2) || depth > func.2 {
+											highest = Some((start, Some(def), depth));
 										},
-										None => if depth + depth2 > func.2 {
-											highest = Some((start, Some(def), depth + depth2));
+										None => if depth > func.2 {
+											highest = Some((start, Some(def), depth));
 										}
 									},
-									None => highest = Some((start, Some(def), depth + depth2))
+									None => highest = Some((start, Some(def), depth))
 								}
 							} else {
 								let mut j = 1;
@@ -680,14 +681,14 @@ pub fn parse_statement(tokens: &Vec<Token>, functions: &Vec<Function>, i: &mut u
 									if let Some(def) = is_defined(functions, &name[..name.len() - j]) {
 										match highest {
 											Some(func) => match func.1 {
-												Some(def2) => if (def.precedence > def2.precedence && depth + depth2 == func.2) || depth + depth2 > func.2 {
-													highest = Some((start, Some(def), depth + depth2));
+												Some(def2) => if (def.precedence > def2.precedence && depth == func.2) || depth > func.2 {
+													highest = Some((start, Some(def), depth));
 												},
-												None => if depth + depth2 > func.2 {
-													highest = Some((start, Some(def), depth + depth2));
+												None => if depth > func.2 {
+													highest = Some((start, Some(def), depth));
 												}
 											},
-											None => highest = Some((start, Some(def), depth + depth2))
+											None => highest = Some((start, Some(def), depth))
 										}
 										
 										break;
@@ -863,11 +864,12 @@ pub fn parse2(tokens: &Vec<Token>, functions: &Vec<Function>, i: &mut usize) {
 				if let Kind::GroupOp(ref op) = tokens[*i].kind {
 					if op == "{" {
 						nests += 1;
-						if let Some(token) = parse_statement(tokens, functions, i) {
-							body.push(token);
-						} else {
-							body.push(start); // Should this really be pushing start instead of *i?
-						}
+						
+						body.push(*i);
+						parse2(tokens, functions, i);
+						
+						*i += 1;
+						continue;
 					}
 				}
 				
@@ -956,7 +958,7 @@ fn del_all_children(tokens: &mut Vec<Token>, children: &Vec<usize>) {
 	}
 }
 
-fn move_children(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<Token>, parent: usize) {
+fn add_to_code(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<Token>, parent: usize) {
 	match tokens[parent].kind {
 		Kind::Var(ref name, _) => if let Some(def) = is_defined(functions, &name) {
 			let new_parent = tokens[parent].clone();
@@ -964,7 +966,7 @@ fn move_children(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<
 			
 			let mut i = 0;
 			while i < def.pos {
-				move_children(tokens, functions, code, children[i]);
+				add_to_code(tokens, functions, code, children[i]);
 				
 				i += 1;
 			}
@@ -974,7 +976,7 @@ fn move_children(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<
 			
 			i += 1;
 			while i < tokens.len() && i < def.args.len() + 1 {
-				move_children(tokens, functions, code, children[i - 1]);
+				add_to_code(tokens, functions, code, children[i - 1]);
 				
 				i += 1;
 			}
@@ -1006,7 +1008,7 @@ fn move_children(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<
 				
 				let mut i = 0;
 				while i < def.pos {
-					move_children(tokens, functions, code, children[i]);
+					add_to_code(tokens, functions, code, children[i]);
 					
 					i += 1;
 				}
@@ -1016,7 +1018,7 @@ fn move_children(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<
 				
 				i += 1;
 				while i < tokens.len() && i < def.args.len() + 1 {
-					move_children(tokens, functions, code, children[i - 1]);
+					add_to_code(tokens, functions, code, children[i - 1]);
 					
 					i += 1;
 				}
@@ -1025,7 +1027,20 @@ fn move_children(tokens: &Vec<Token>, functions: &Vec<Function>, code: &mut Vec<
 			}
 		},
 		
-		Kind::GroupOp(ref op) if op == "{" => (), // WIP
+		Kind::GroupOp(ref op) if op == "{" => {
+			let new_parent = tokens[parent].clone();
+			let mut children = tokens[parent].children.borrow_mut();
+			
+			new_parent.children.borrow_mut().clear();
+			code.push(new_parent);
+			
+			let mut i = 0;
+			while i < tokens.len() && i < children.len() {
+				add_to_code(tokens, functions, code, children[i]);
+				
+				i += 1;
+			}
+		},
 		
 		_ => code.push(tokens[parent].clone())
 	}
@@ -1046,7 +1061,7 @@ pub fn parse3(tokens: &mut Vec<Token>, macro_funcs: &mut Vec<MacroFunction>, fun
 						for (a, arg) in args.iter().enumerate() {
 							for token in macro_funcs[j].code.iter() {
 								match token.kind {
-									Kind::Var(ref name, _) if name == &macro_funcs[j].func.args[a].name => move_children(tokens, functions, &mut new_code, *arg),
+									Kind::Var(ref name, _) if name == &macro_funcs[j].func.args[a].name => add_to_code(tokens, functions, &mut new_code, *arg),
 									_ => new_code.push(token.clone())
 								}
 							}
@@ -1055,7 +1070,7 @@ pub fn parse3(tokens: &mut Vec<Token>, macro_funcs: &mut Vec<MacroFunction>, fun
 								new_points.push(Vec::new());
 								for token in point.iter() {
 									match token.kind {
-										Kind::Var(ref name, _) if name == &macro_funcs[j].func.args[a].name => move_children(tokens, functions, &mut new_points[p], *arg),
+										Kind::Var(ref name, _) if name == &macro_funcs[j].func.args[a].name => add_to_code(tokens, functions, &mut new_points[p], *arg),
 										_ => new_points[p].push(token.clone())
 									}
 								}
@@ -1125,11 +1140,11 @@ pub fn parse3(tokens: &mut Vec<Token>, macro_funcs: &mut Vec<MacroFunction>, fun
 							.expect("failed to compile Rust code");
 					
 					if out.stdout.len() > 0 {
-						print!("{}", str::from_utf8(&out.stdout).unwrap());
+						println!("{}", str::from_utf8(&out.stdout).unwrap());
 					}
 					
 					if out.stderr.len() > 0 {
-						print!("{}", str::from_utf8(&out.stderr).unwrap());
+						println!("{}", str::from_utf8(&out.stderr).unwrap());
 						error = true;
 					}
 					
@@ -1147,7 +1162,7 @@ pub fn parse3(tokens: &mut Vec<Token>, macro_funcs: &mut Vec<MacroFunction>, fun
 						};
 						
 						if out.stdout.len() > 0 {
-							print!("{}", str::from_utf8(&out.stdout).unwrap());
+							println!("{}", str::from_utf8(&out.stdout).unwrap());
 							io::stdout().flush()?;
 						}
 						
@@ -1191,7 +1206,7 @@ pub fn parse3(tokens: &mut Vec<Token>, macro_funcs: &mut Vec<MacroFunction>, fun
 									}
 								}
 							} else {
-								print!("{}", str::from_utf8(&out.stderr).unwrap());
+								println!("{}", str::from_utf8(&out.stderr).unwrap());
 							}
 						} else {
 							tokens.insert(*i, Token {
@@ -1216,8 +1231,14 @@ pub fn parse3(tokens: &mut Vec<Token>, macro_funcs: &mut Vec<MacroFunction>, fun
 					//////// DELETE CREATED FILES ////////
 					
 					fs::remove_file("macros\\macro.rs")?;
-					fs::remove_file("macros\\macro.exe")?;
-					fs::remove_file("macros\\macro.pdb")?;
+					
+					if !error {
+						fs::remove_file("macros\\macro.exe")?;
+						fs::remove_file("macros\\macro.pdb")?;
+					} else {
+						return Err(Error::new(ErrorKind::Other, "compilation of macro failed"));
+					}
+					
 //					fs::remove_dir("macros")?; // Doesn't work (on Windows) for some reason?
 					
 					break;
